refactoring: extract gmailutils.NormalizeAndSplit() + tests
